+++ packages/stream_video_push_notification/android/src/main/kotlin/io/getstream/video/flutter/stream_video_push_notification/TransparentActivity.kt
@@ -1,0 +1,45 @@
+package io.getstream.video.flutter.stream_video_push_notification
+
+import android.app.Activity
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+
+class TransparentActivity : Activity() {
+
+    companion object {
+        fun getIntent(context: Context, action: String, data: Bundle?): Intent {
+            val intent = Intent(context, TransparentActivity::class.java)
+            intent.action = action
+            intent.putExtra("data", data)
+            intent.addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION)
+            intent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY)
+            intent.addFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT)
+            intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+            intent.addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP)
+            return intent
+        }
+    }
+
+
+    override fun onStart() {
+        super.onStart()
+        setVisible(false)
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        val data = intent.getBundleExtra("data")
+
+        val broadcastIntent = IncomingCallBroadcastReceiver.getIntent(this, intent.action!!, data)
+        broadcastIntent.addFlags(Intent.FLAG_RECEIVER_FOREGROUND)
+        sendBroadcast(broadcastIntent)
+
+        val activityIntent = AppUtils.getAppIntent(this, intent.action, data)
+        startActivity(activityIntent)
+
+        finish()
+        overridePendingTransition(0, 0)
+    }
+}